Split Jetty instrumentation out and enable by default
Prior jetty instrumentation used a different operation name (`jetty.request`), but this would impact metrics if enabled when `servlet.request` was previously expected. Since it was disabled by default and considered "beta", this is deemed an acceptable change.

Current instrumentation should work with Jetty 7-10 (though 10 isn't explicitly tested because it wasn't compiled for Java 8).  Jetty 11 is still not supported since it requires Servlet 4.

Note the odd overlap in supported versions:
* Jetty 8.0 is instrumented by 7.0 instrumentation
* Jetty 8.1-8.2 is instrumented by 7.6 instrumentation

[skip ci]